Update WRF sample
We can now use YAML instead of JSON, which is much nicer for jobs
with complicated scripts. There's no longer a need to redirect to
runsim.sh because the contents of runsim.sh can comfortably live
in the job configuration file itself.

The user no longer needs to create an instance template and point
the Batch job at that template, in order to get a boot image based
on the HPC CentOS 7 image. Instead the job spec can use the boot
image shortcut "batch-hpc-centos" to get an image based on
HPC CentOS 7, and with the Batch agent preinstalled.

Other minor cleanup, rewording.