CNN training sample
The sample shows how to train a CNN using a mulit-node Batch job.
The sample model can be trained to distinguish images of cats from
images of dogs using the Kaggle dogs-vs-cats data set, as described
in the README.

The sample uses two Batch jobs, a large preprocessing job that runs
on general purpose hardware, and a smaller GPU-driven job to train
the model.